refactor(ui): move the chip filter primitives into the shared UI library
`DotChipFilterComponent`, `DotFilterListItemComponent` and the two
`CHIP_FILTER_*_PT` constants lived in `@dotcms/portlets/content-drive/ui`, so
every portlet that wanted a filter chip imported from another portlet's UI lib.
Seven consumers across content-drive, users, publishing-queue and experiments
already did exactly that. They now come from `@dotcms/ui`, where shared,
domain-agnostic UI belongs.

Pure move: no behaviour change. Both components had no content-drive coupling —
the only cross-lib import was `DotMessagePipe`, now relative. Only the two
`CHIP_FILTER_*` constants moved; `HEADER_COLUMNS`, `ALL_FOLDER` and the rest of
content-drive's constants stay put.

One test needed adjusting: `libs/ui` runs `@happy-dom/jest-environment` while
content-drive-ui runs jsdom, and Spectator's `dispatchKeyboardEvent` builds its
event with the legacy `initKeyboardEvent`, which happy-dom does not implement.
The two keydown tests now dispatch a real `KeyboardEvent`, which works under
both.
