+++ core-web/libs/portlets/dot-content-drive/portlet/src/lib/components/dot-content-drive-toolbar/components/dot-content-drive-content-type-filter/dot-content-drive-content-type-filter.component.spec.ts
@@ -22,12 +22,13 @@ import {
     DotCMSContentType,
     StructureTypeView
 } from '@dotcms/dotcms-models';
-import { DotChipFilterComponent } from '@dotcms/portlets/content-drive/ui';
+import { DotChipFilterComponent } from '@dotcms/ui';
 import { MockDotMessageService } from '@dotcms/utils-testing';
 
 import { DotContentDriveContentTypeFilterComponent } from './dot-content-drive-content-type-filter.component';
 
 import { DotContentDriveStore } from '../../../../store/dot-content-drive.store';
+
 
 const BASE_TYPES: StructureTypeView[] = [
     { name: 'CONTENT', label: 'Content', types: [] },
