+++ core-web/libs/portlets/dot-content-drive/portlet/src/lib/components/dot-content-drive-toolbar/components/dot-content-drive-language-field/dot-content-drive-language-field.spec.ts
@@ -14,12 +14,13 @@ import { Popover } from 'primeng/popover';
 
 import { DotLanguagesService, DotMessageService } from '@dotcms/data-access';
 import { DotLanguage } from '@dotcms/dotcms-models';
-import { DotChipFilterComponent } from '@dotcms/portlets/content-drive/ui';
+import { DotChipFilterComponent } from '@dotcms/ui';
 import { createFakeLanguage, MockDotMessageService } from '@dotcms/utils-testing';
 
 import { DotContentDriveLanguageFieldComponent } from './dot-content-drive-language-field.component';
 
 import { DotContentDriveStore } from '../../../../store/dot-content-drive.store';
+
 
 const MOCK_LANGUAGES: DotLanguage[] = [
     createFakeLanguage({
