+++ core-web/libs/portlets/dot-content-drive/ui/src/index.ts
@@ -1,6 +1,4 @@
 export * from './lib/dot-folder-list-view/dot-folder-list-view.component';
 export * from './lib/dot-tree-folder/dot-tree-folder.component';
-export * from './lib/dot-chip-filter/dot-chip-filter.component';
-export * from './lib/dot-filter-list-item/dot-filter-list-item.component';
 export * from './lib/shared/models';
 export * from './lib/shared/constants';
