+++ core-web/libs/portlets/dot-experiments/portlet/src/lib/components/dot-experiment-status-filter/dot-experiment-status-filter.component.ts
@@ -19,9 +19,9 @@ import {
     CHIP_FILTER_LISTBOX_PT,
     CHIP_FILTER_POPOVER_PT,
     DotChipFilterComponent,
-    DotFilterListItemComponent
-} from '@dotcms/portlets/content-drive/ui';
-import { DotMessagePipe } from '@dotcms/ui';
+    DotFilterListItemComponent,
+    DotMessagePipe
+} from '@dotcms/ui';
 
 /** A single status entry rendered inside the filter listbox. */
 interface StatusFilterOption {
