+++ core-web/libs/ui/src/lib/components/dot-chip-filter/dot-chip-filter.component.html
@@ -7,7 +7,7 @@
 @if (active()) {
     <button
         type="button"
-        class="flex items-center justify-center bg-transparent border-0 p-0 cursor-pointer text-current"
+        class="flex cursor-pointer items-center justify-center border-0 bg-transparent p-0 text-current"
         [attr.aria-label]="'dot.common.remove' | dm"
         (click)="onRemove($event)"
         data-testid="chip-remove">
